requirements.txt
telethon
pillow
datetime